chore: move repo/sqlc_v2 to repo/sqlc
