+++ internal/repo/sqlc/repository_test.go
@@ -87,7 +87,7 @@ func Test_NewRepo(t *testing.T) {
 			name:   "works",
 			site:   "test",
 			db:     nil,
-			expect: &SqlcRepo{site: "test", db: nil},
+			expect: &SqlcRepo{db: nil},
 		},
 	}
 
@@ -95,7 +95,7 @@ func Test_NewRepo(t *testing.T) {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
 			t.Parallel()
-			result := NewRepo(test.site, test.db)
+			result := NewRepo(test.db)
 			if test.db != result.db {
 				t.Errorf("got: %v, want: %v", result, test.expect)
 			}
@@ -114,6 +114,7 @@ func TestSqlcRepo_CreateBook(t *testing.T) {
 
 	t.Cleanup(func() {
 		db.Exec("delete from books where site=$1", site)
+
 		db.Close()
 	})
 
@@ -127,21 +128,21 @@ func TestSqlcRepo_CreateBook(t *testing.T) {
 	}{
 		{
 			name:       "create book with new id to hash code 0",
-			r:          NewRepo(site, db),
+			r:          NewRepo(db),
 			bk:         model.Book{Site: site, ID: 1, HashCode: 100, Writer: model.Writer{ID: 10}},
 			expectBook: model.Book{Site: site, ID: 1, HashCode: 0, Writer: model.Writer{ID: 10}},
 			expectErr:  false,
 		},
 		{
 			name:       "create book with existing id with input hash code",
-			r:          NewRepo(site, db),
+			r:          NewRepo(db),
 			bk:         model.Book{Site: site, ID: 1, HashCode: 100, Writer: model.Writer{ID: 10}},
 			expectBook: model.Book{Site: site, ID: 1, HashCode: 100, Writer: model.Writer{ID: 10}},
 			expectErr:  false,
 		},
 		{
 			name:       "fail to create book with existing id and hash",
-			r:          NewRepo(site, db),
+			r:          NewRepo(db),
 			bk:         model.Book{Site: site, ID: 1, HashCode: 100, Writer: model.Writer{ID: 10}},
 			expectBook: model.Book{Site: site, ID: 1, HashCode: 100, Writer: model.Writer{ID: 10}},
 			expectErr:  true,
@@ -161,7 +162,7 @@ func TestSqlcRepo_CreateBook(t *testing.T) {
 			})
 
 			t.Run("book in db", func(t *testing.T) {
-				bk, err := test.r.FindBookByIdHash(context.Background(), test.bk.ID, test.bk.HashCode)
+				bk, err := test.r.FindBookByIdHash(context.Background(), site, test.bk.ID, test.bk.HashCode)
 				if err != nil {
 					t.Fatalf("query got error: %v", err)
 				}
@@ -189,7 +190,7 @@ func TestSqlcRepo_UpdateBook(t *testing.T) {
 		db.Close()
 	})
 
-	bksDB := stubData(t, NewRepo(site, db), site)
+	bksDB := stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
@@ -202,7 +203,7 @@ func TestSqlcRepo_UpdateBook(t *testing.T) {
 	}{
 		{
 			name: "update not existing book",
-			r:    NewRepo(site, db),
+			r:    NewRepo(db),
 			inputBook: &model.Book{
 				Site: site, ID: -1, HashCode: 0, Title: "hello",
 			},
@@ -212,7 +213,7 @@ func TestSqlcRepo_UpdateBook(t *testing.T) {
 		},
 		{
 			name: "update error book to in progress without changing writer id and error",
-			r:    NewRepo(site, db),
+			r:    NewRepo(db),
 			inputBook: &model.Book{
 				Site: bksDB[4].Site, ID: bksDB[4].ID, HashCode: bksDB[4].HashCode,
 				Title: "t", Writer: model.Writer{Name: bksDB[0].Writer.Name}, Type: "t",
@@ -241,7 +242,7 @@ func TestSqlcRepo_UpdateBook(t *testing.T) {
 			})
 
 			t.Run("book in db", func(t *testing.T) {
-				bk, err := test.r.FindBookByIdHash(context.Background(), test.inputBook.ID, test.inputBook.HashCode)
+				bk, err := test.r.FindBookByIdHash(context.Background(), site, test.inputBook.ID, test.inputBook.HashCode)
 				if (err != nil) != test.expectQueryErr {
 					t.Errorf("query got error: %v; want error: %v", err, test.expectQueryErr)
 				}
@@ -266,7 +267,7 @@ func TestSqlcRepo_FindBookByID(t *testing.T) {
 		db.Close()
 	})
 
-	bksDB := stubData(t, NewRepo(site, db), site)
+	bksDB := stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
@@ -279,7 +280,7 @@ func TestSqlcRepo_FindBookByID(t *testing.T) {
 	}{
 		{
 			name:         "find not existing book",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			id:           0,
 			expectResult: nil,
 			expectHash:   0,
@@ -287,7 +288,7 @@ func TestSqlcRepo_FindBookByID(t *testing.T) {
 		},
 		{
 			name:         "find book with largest id",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			id:           2,
 			expectResult: &bksDB[2],
 			expectHash:   100,
@@ -298,7 +299,7 @@ func TestSqlcRepo_FindBookByID(t *testing.T) {
 	for _, test := range tests {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
-			result, err := test.r.FindBookById(context.Background(), test.id)
+			result, err := test.r.FindBookById(context.Background(), site, test.id)
 			t.Run("result", func(t *testing.T) {
 				if (err != nil) != test.expectErr {
 					t.Errorf("got error: %v; want error: %v", err, test.expectErr)
@@ -325,7 +326,7 @@ func TestSqlcRepo_FindBookByIDHash(t *testing.T) {
 		db.Close()
 	})
 
-	bksDB := stubData(t, NewRepo(site, db), site)
+	bksDB := stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
@@ -338,7 +339,7 @@ func TestSqlcRepo_FindBookByIDHash(t *testing.T) {
 	}{
 		{
 			name:         "find not existing book",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			id:           0,
 			hashcode:     0,
 			expectResult: nil,
@@ -346,7 +347,7 @@ func TestSqlcRepo_FindBookByIDHash(t *testing.T) {
 		},
 		{
 			name:         "find book with correct id hash",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			id:           2,
 			hashcode:     0,
 			expectResult: &bksDB[1],
@@ -357,7 +358,7 @@ func TestSqlcRepo_FindBookByIDHash(t *testing.T) {
 	for _, test := range tests {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
-			result, err := test.r.FindBookByIdHash(context.Background(), test.id, test.hashcode)
+			result, err := test.r.FindBookByIdHash(context.Background(), site, test.id, test.hashcode)
 			t.Run("result", func(t *testing.T) {
 				if (err != nil) != test.expectErr {
 					t.Errorf("got error: %v; want error: %v", err, test.expectErr)
@@ -389,24 +390,27 @@ func TestSqlcRepo_FindAllBooks(t *testing.T) {
 		db.Close()
 	})
 
-	bksDB := stubData(t, NewRepo(site, db), site)
+	bksDB := stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
 		name         string
 		r            repo.Repository
+		site         string
 		expectResult []model.Book
 		expectErr    bool
 	}{
 		{
 			name:         "works",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
+			site:         site,
 			expectResult: bksDB,
 			expectErr:    false,
 		},
 		{
 			name:         "works for empty",
-			r:            NewRepo("empty", db),
+			r:            NewRepo(db),
+			site:         "empty",
 			expectResult: nil,
 			expectErr:    false,
 		},
@@ -415,7 +419,7 @@ func TestSqlcRepo_FindAllBooks(t *testing.T) {
 	for _, test := range tests {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
-			result, err := test.r.FindAllBooks(context.Background())
+			result, err := test.r.FindAllBooks(context.Background(), test.site)
 			if (err != nil) != test.expectErr {
 				t.Errorf("got error: %v; want error: %v", err, test.expectErr)
 			}
@@ -445,7 +449,7 @@ func TestSqlcRepo_FindBooksForUpdate(t *testing.T) {
 		db.Close()
 	})
 
-	bksDB := stubData(t, NewRepo(site, db), site)
+	bksDB := stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
@@ -456,7 +460,7 @@ func TestSqlcRepo_FindBooksForUpdate(t *testing.T) {
 	}{
 		{
 			name:         "works",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			expectResult: []model.Book{bksDB[4], bksDB[3], bksDB[2], bksDB[0]},
 			expectErr:    false,
 		},
@@ -465,7 +469,7 @@ func TestSqlcRepo_FindBooksForUpdate(t *testing.T) {
 	for _, test := range tests {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
-			result, err := test.r.FindBooksForUpdate(context.Background())
+			result, err := test.r.FindBooksForUpdate(context.Background(), site)
 			if (err != nil) != test.expectErr {
 				t.Errorf("got error: %v; want error: %v", err, test.expectErr)
 			}
@@ -495,7 +499,7 @@ func TestSqlcRepo_FindBooksForDownload(t *testing.T) {
 		db.Close()
 	})
 
-	bksDB := stubData(t, NewRepo(site, db), site)
+	bksDB := stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
@@ -506,7 +510,7 @@ func TestSqlcRepo_FindBooksForDownload(t *testing.T) {
 	}{
 		{
 			name:         "works",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			expectResult: []model.Book{bksDB[2]},
 			expectErr:    false,
 		},
@@ -515,7 +519,7 @@ func TestSqlcRepo_FindBooksForDownload(t *testing.T) {
 	for _, test := range tests {
 		test := test
 		t.Run(test.name, func(t *testing.T) {
-			result, err := test.r.FindBooksForDownload(context.Background())
+			result, err := test.r.FindBooksForDownload(context.Background(), site)
 			if (err != nil) != test.expectErr {
 				t.Errorf("got error: %v; want error: %v", err, test.expectErr)
 			}
@@ -536,16 +540,22 @@ func TestSqlcRepo_FindBooksByTitleWriter(t *testing.T) {
 		t.FailNow()
 	}
 	site := "bk_tit_wrt/find"
+	siteV2 := "bk_tit_wrt/v2"
 
 	t.Cleanup(func() {
 		db.Exec("delete from books where site=$1", site)
 		db.Exec("delete from writers where id>0 and name like $1", site+"%")
 		db.Exec("delete from errors where site=$1", site)
 
+		db.Exec("delete from books where site=$1", siteV2)
+		db.Exec("delete from writers where id>0 and name like $1", siteV2+"%")
+		db.Exec("delete from errors where site=$1", siteV2)
+
 		db.Close()
 	})
 
-	bksDB := stubData(t, NewRepo(site, db), site)
+	bksDB := stubData(t, NewRepo(db), site)
+	bksDBV2 := stubData(t, NewRepo(db), siteV2)
 
 	t.Parallel()
 	tests := []struct {
@@ -559,43 +569,48 @@ func TestSqlcRepo_FindBooksByTitleWriter(t *testing.T) {
 		expectErr    bool
 	}{
 		{
-			name:         "works",
-			r:            NewRepo(site, db),
-			title:        "title",
-			writer:       "writer",
-			limit:        10,
-			offset:       0,
-			expectResult: []model.Book{bksDB[3], bksDB[2], bksDB[1], bksDB[0]},
-			expectErr:    false,
+			name:   "works",
+			r:      NewRepo(db),
+			title:  "title",
+			writer: "writer",
+			limit:  10,
+			offset: 0,
+			expectResult: []model.Book{
+				bksDBV2[3], bksDB[3],
+				bksDBV2[2], bksDB[2],
+				bksDBV2[1], bksDB[1],
+				bksDBV2[0], bksDB[0],
+			},
+			expectErr: false,
 		},
 		{
 			name:         "works with limit",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			title:        "title",
 			writer:       "writer",
 			limit:        1,
 			offset:       0,
-			expectResult: []model.Book{bksDB[3]},
+			expectResult: []model.Book{bksDBV2[3]},
 			expectErr:    false,
 		},
 		{
 			name:         "works with offset",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			title:        "title",
 			writer:       "writer",
 			limit:        1,
 			offset:       1,
-			expectResult: []model.Book{bksDB[2]},
+			expectResult: []model.Book{bksDB[3]},
 			expectErr:    false,
 		},
 		{
 			name:         "return all books match either title or writer",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			title:        "title 1",
 			writer:       "writer 3",
 			limit:        5,
 			offset:       0,
-			expectResult: []model.Book{bksDB[3], bksDB[0]},
+			expectResult: []model.Book{bksDBV2[3], bksDB[3], bksDBV2[0], bksDB[0]},
 			expectErr:    false,
 		},
 	}
@@ -617,17 +632,25 @@ func TestSqlcRepo_FindBooksByRandom(t *testing.T) {
 	if !assert.NoError(t, err, "Failed to open database") {
 		t.FailNow()
 	}
-	site := "random_bk/find"
+	site := "rand_bk/find"
+	siteV2 := "rand_bk/find_v2"
 
 	t.Cleanup(func() {
 		db.Exec("delete from books where site=$1", site)
 		db.Exec("delete from writers where id>0 and name like $1", site+"%")
 		db.Exec("delete from errors where site=$1", site)
 
+		db.Exec("delete from books where site=$1", siteV2)
+		db.Exec("delete from writers where id>0 and name like $1", siteV2+"%")
+		db.Exec("delete from errors where site=$1", siteV2)
+
 		db.Close()
 	})
+	t.Cleanup(func() {
+	})
 
-	stubData(t, NewRepo(site, db), site)
+	stubData(t, NewRepo(db), site)
+	stubData(t, NewRepo(db), siteV2)
 
 	t.Parallel()
 	tests := []struct {
@@ -639,9 +662,9 @@ func TestSqlcRepo_FindBooksByRandom(t *testing.T) {
 	}{
 		{
 			name:         "works",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			limit:        10,
-			expectLength: 2,
+			expectLength: 4,
 			expectErr:    false,
 		},
 	}
@@ -655,6 +678,11 @@ func TestSqlcRepo_FindBooksByRandom(t *testing.T) {
 			}
 			if len(result) != test.expectLength {
 				t.Errorf("query got:  %v\nwant length: %v", result, test.expectLength)
+			}
+			for _, bk := range result {
+				if bk.Site != site && bk.Site != siteV2 {
+					t.Errorf("query got book with unexpected site: %v", bk.Site)
+				}
 			}
 		})
 	}
@@ -675,9 +703,9 @@ func TestSqlcRepo_UpdateBooksStatus(t *testing.T) {
 		db.Close()
 	})
 
-	r := NewRepo(site, db)
+	r := NewRepo(db)
 
-	bksDB := stubData(t, NewRepo(site, db), site)
+	bksDB := stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
@@ -690,7 +718,7 @@ func TestSqlcRepo_UpdateBooksStatus(t *testing.T) {
 	}{
 		{
 			name:      "works for updating in progress to end",
-			r:         NewRepo(site, db),
+			r:         NewRepo(db),
 			bkID:      3,
 			bkHash:    0,
 			expectErr: false,
@@ -703,7 +731,7 @@ func TestSqlcRepo_UpdateBooksStatus(t *testing.T) {
 		},
 		{
 			name:      "works for updating download to false",
-			r:         NewRepo(site, db),
+			r:         NewRepo(db),
 			bkID:      1,
 			bkHash:    0,
 			expectErr: false,
@@ -716,7 +744,7 @@ func TestSqlcRepo_UpdateBooksStatus(t *testing.T) {
 		},
 		{
 			name:      "works for not updating in progress to end",
-			r:         NewRepo(site, db),
+			r:         NewRepo(db),
 			bkID:      2,
 			bkHash:    100,
 			expectErr: false,
@@ -741,13 +769,12 @@ func TestSqlcRepo_UpdateBooksStatus(t *testing.T) {
 			r.UpdateBook(context.Background(), &bksDB[0])
 
 			t.Parallel()
-
 			err := test.r.UpdateBooksStatus(context.Background())
 			if (err != nil) != test.expectErr {
 				t.Errorf("got error: %v; want error: %v", err, test.expectErr)
 			}
 
-			bk, err := test.r.FindBookByIdHash(context.Background(), test.bkID, test.bkHash)
+			bk, err := test.r.FindBookByIdHash(context.Background(), site, test.bkID, test.bkHash)
 			if err != nil {
 				t.Errorf("book fail to fetch: id: %v; hash: %v; err: %v", test.bkID, test.bkHash, err)
 				return
@@ -772,9 +799,9 @@ func BenchmarkSqlcRepo_UpdateBooksStatus(b *testing.B) {
 		db.Close()
 	})
 
-	r := NewRepo(site, db)
+	r := NewRepo(db)
 
-	stubData(b, NewRepo(site, db), site)
+	stubData(b, NewRepo(db), site)
 
 	for n := 0; n < b.N; n++ {
 		r.UpdateBooksStatus(context.Background())
@@ -794,9 +821,9 @@ func TestSqlcRepo_FindAllBookIDs(t *testing.T) {
 		db.Close()
 	})
 
-	r := NewRepo(site, db)
+	r := NewRepo(db)
 
-	stubData(t, NewRepo(site, db), site)
+	stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
@@ -818,7 +845,7 @@ func TestSqlcRepo_FindAllBookIDs(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			t.Parallel()
 
-			got, err := test.r.FindAllBookIDs(context.Background())
+			got, err := test.r.FindAllBookIDs(context.Background(), site)
 			assert.Equal(t, test.want, got)
 			assert.ErrorIs(t, test.wantError, err)
 		})
@@ -840,7 +867,7 @@ func TestSqlcRepo_SaveWriter(t *testing.T) {
 		db.Close()
 	})
 
-	bksDB := stubData(t, NewRepo(site, db), site)
+	bksDB := stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
@@ -851,13 +878,13 @@ func TestSqlcRepo_SaveWriter(t *testing.T) {
 	}{
 		{
 			name:      "save existing writer",
-			r:         NewRepo(site, db),
+			r:         NewRepo(db),
 			writer:    &model.Writer{ID: 0, Name: bksDB[0].Writer.Name},
 			expectErr: false,
 		},
 		{
 			name:      "save new writer",
-			r:         NewRepo(site, db),
+			r:         NewRepo(db),
 			writer:    &model.Writer{ID: 0, Name: site + " new writer"},
 			expectErr: false,
 		},
@@ -892,7 +919,7 @@ func TestSqlcRepo_SaveError(t *testing.T) {
 		db.Close()
 	})
 
-	stubData(t, NewRepo(site, db), site)
+	stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
@@ -905,7 +932,7 @@ func TestSqlcRepo_SaveError(t *testing.T) {
 	}{
 		{
 			name:         "create error for existing book",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			bk:           &model.Book{Site: site, ID: 1},
 			e:            errors.New("create error"),
 			expectErrStr: "create error",
@@ -913,7 +940,7 @@ func TestSqlcRepo_SaveError(t *testing.T) {
 		},
 		{
 			name:         "update error for existing book",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			bk:           &model.Book{Site: site, ID: 1},
 			e:            errors.New("update error"),
 			expectErrStr: "update error",
@@ -921,7 +948,7 @@ func TestSqlcRepo_SaveError(t *testing.T) {
 		},
 		{
 			name:         "delete error for existing book",
-			r:            NewRepo(site, db),
+			r:            NewRepo(db),
 			bk:           &model.Book{Site: site, ID: 1},
 			e:            nil,
 			expectErrStr: "",
@@ -939,7 +966,7 @@ func TestSqlcRepo_SaveError(t *testing.T) {
 			})
 
 			t.Run("error in db", func(t *testing.T) {
-				bk, err := test.r.FindBookByIdHash(context.Background(), test.bk.ID, test.bk.HashCode)
+				bk, err := test.r.FindBookByIdHash(context.Background(), site, test.bk.ID, test.bk.HashCode)
 				if err != nil {
 					t.Errorf("query got error: %v; want error: %v", err, false)
 				}
@@ -954,10 +981,8 @@ func TestSqlcRepo_SaveError(t *testing.T) {
 }
 
 // func TestSqlcRepo_Backup(t *testing.T) {
-// db, err := OpenDatabaseByConfig(conf)
-// if !assert.NoError(t, err, "Failed to open database") {
-// 	t.FailNow()
-// }
+// 	t.Parallel()
+// 	StubPsqlConn()
 // 	db := testDB
 // 	site := "backup"
 
@@ -966,10 +991,9 @@ func TestSqlcRepo_SaveError(t *testing.T) {
 // 		db.Exec("delete from writers where id>0 and name like $1", site+"%")
 // 		db.Exec("delete from errors where site=$1", site)
 
-// db.Close()
 // 	})
 
-// 	stubData(t, NewRepo(site, db), site)
+// 	stubData(t, NewRepo(db), site)
 
 // 	tests := []struct {
 // 		name      string
@@ -979,7 +1003,7 @@ func TestSqlcRepo_SaveError(t *testing.T) {
 // 	}{
 // 		{
 // 			name:      "works",
-// 			r:         NewRepo(site, db),
+// 			r:         NewRepo(db),
 // 			path:      "/" + site,
 // 			expectErr: false,
 // 		},
@@ -988,7 +1012,7 @@ func TestSqlcRepo_SaveError(t *testing.T) {
 // 	for _, test := range tests {
 // 		test := test
 // 		t.Run(test.name, func(t *testing.T) {
-// 			err := test.r.Backup(context.Background(), test.path)
+// 			err := test.r.Backup(context.Background(), site, test.path)
 // 			if (err != nil) != test.expectErr {
 // 				t.Errorf("got err: %v; want err: %v", err, test.expectErr)
 // 			}
@@ -1011,7 +1035,7 @@ func TestSqlcRepo_Stats(t *testing.T) {
 		db.Close()
 	})
 
-	stubData(t, NewRepo(site, db), site)
+	stubData(t, NewRepo(db), site)
 
 	t.Parallel()
 	tests := []struct {
@@ -1021,7 +1045,7 @@ func TestSqlcRepo_Stats(t *testing.T) {
 	}{
 		{
 			name: "works",
-			r:    NewRepo(site, db),
+			r:    NewRepo(db),
 			expect: repo.Summary{
 				BookCount: 5, WriterCount: 5, ErrorCount: 1, DownloadCount: 2,
 				UniqueBookCount: 4, MaxBookID: 4,
@@ -1039,7 +1063,7 @@ func TestSqlcRepo_Stats(t *testing.T) {
 		t.Run(test.name, func(t *testing.T) {
 			t.Parallel()
 
-			result := test.r.Stats(context.Background())
+			result := test.r.Stats(context.Background(), site)
 			assert.Equal(t, test.expect, result)
 		})
 	}
